+++ nodes/persistence-node/src/types/NotFoundError.ts
